remove the config and rewire a lot of plumbing
