+++ pkg/certificate/providers/certmanager/types.go
@@ -2,7 +2,6 @@
 package certmanager
 
 import (
-	"sync"
 	"time"
 
 	cmmeta "github.com/jetstack/cert-manager/pkg/apis/meta/v1"
@@ -10,9 +9,7 @@ import (
 	cmlisters "github.com/jetstack/cert-manager/pkg/client/listers/certmanager/v1"
 
 	"github.com/openservicemesh/osm/pkg/certificate"
-	"github.com/openservicemesh/osm/pkg/configurator"
 	"github.com/openservicemesh/osm/pkg/logger"
-	"github.com/openservicemesh/osm/pkg/messaging"
 )
 
 const (
@@ -26,15 +23,10 @@ var (
 )
 
 // CertManager implements certificate.Manager
-type CertManager struct {
+type Provider struct {
 	// The Certificate Authority root certificate to be used by this certificate
 	// manager.
 	ca *certificate.Certificate
-
-	// cache holds a local cache of issued certificates as
-	// *certificate.Certificates
-	cache     map[certificate.CommonName]*certificate.Certificate
-	cacheLock sync.RWMutex
 
 	// Control plane namespace where CertificateRequests are created.
 	namespace string
@@ -48,11 +40,13 @@ type CertManager struct {
 	// crLister is used to list CertificateRequests in the given namespace.
 	crLister cmlisters.CertificateRequestNamespaceLister
 
-	cfg configurator.Configurator
-
 	// Issuing certificate properties.
-	serviceCertValidityDuration time.Duration
-	keySize                     int
+	keySize int
+}
 
-	msgBroker *messaging.Broker
+// CertManagerOptions is a type that specifies 'cert-manager.io' certificate provider options
+type Options struct {
+	IssuerName  string
+	IssuerKind  string
+	IssuerGroup string
 }
