+++ pkg/certificate/providers/tresor/errors.go
@@ -8,4 +8,4 @@ var errCreateCert = errors.New("create cert")
 var errGeneratingSerialNumber = errors.New("generate serial number")
 var errGeneratingPrivateKey = errors.New("generate private")
 var errNoIssuingCA = errors.New("no issuing CA")
-var errCertNotFound = errors.New("certificate not found")
+var errNoPrivateKey = errors.New("no private key provided, required for tresor")
